fix(manifest-ui): fix circular dependencies across all categories
Created types.ts files for each category:
- list/types.ts (Product interface)
- messaging/types.ts (ChatMessage interface)
- payment/types.ts (OrderItem interface)
- selection/types.ts (Option interface)

Updated demo/data.ts files to import from ../types:
- list/demo/data.ts
- messaging/demo/data.ts
- payment/demo/data.ts
- selection/demo/data.ts

Updated components to import from ./types and re-export:
- product-list.tsx
- chat-conversation.tsx
- order-summary.tsx
- option-list.tsx

This prevents circular dependencies:
component.tsx -> demo/data.ts -> types.ts (no cycle)
