+++ packages/manifest-ui/registry/messaging/chat-conversation.tsx
@@ -2,30 +2,10 @@
 
 import { ImageMessageBubble, MessageBubble } from './message-bubble'
 
-/**
- * Represents a chat message.
- * @interface ChatMessage
- * @property {"text" | "image"} [type] - Message type (defaults to "text")
- * @property {string} [content] - Message content or image caption
- * @property {string} [image] - Image URL for image messages
- * @property {string} [author] - Sender's name
- * @property {string} [avatarUrl] - Sender's avatar image URL
- * @property {string} [avatarFallback] - Fallback letter for avatar
- * @property {string} [time] - Message timestamp
- * @property {boolean} [isOwn] - Whether message is from current user (defaults to false)
- * @property {"sent" | "delivered" | "read"} [status] - Message delivery status
- */
-export interface ChatMessage {
-  type?: 'text' | 'image'
-  content?: string
-  image?: string
-  author?: string
-  avatarUrl?: string
-  avatarFallback?: string
-  time?: string
-  isOwn?: boolean
-  status?: 'sent' | 'delivered' | 'read'
-}
+// Import types from shared types file to avoid circular dependencies
+import type { ChatMessage } from './types'
+// Re-export for backward compatibility
+export type { ChatMessage } from './types'
 
 /**
  * ═══════════════════════════════════════════════════════════════════════════
