+++ packages/manifest-ui/registry/messaging/types.ts
@@ -1,0 +1,17 @@
+// Shared types for Messaging category components
+
+/**
+ * Represents a chat message with content and metadata.
+ * @interface ChatMessage
+ */
+export interface ChatMessage {
+  type?: 'text' | 'image'
+  content?: string
+  image?: string
+  author?: string
+  avatarUrl?: string
+  avatarFallback?: string
+  time?: string
+  isOwn?: boolean
+  status?: 'sent' | 'delivered' | 'read'
+}
