+++ packages/manifest-ui/registry/payment/types.ts
@@ -1,0 +1,13 @@
+// Shared types for Payment category components
+
+/**
+ * Represents an item in an order.
+ * @interface OrderItem
+ */
+export interface OrderItem {
+  id: string
+  name?: string
+  quantity?: number
+  price?: number
+  image?: string
+}
